+++ pyremo/cli.py
@@ -5,7 +5,7 @@ import sys
 
 from dask.distributed import Client
 
-from .prsint import cli as prsint_cli
+from .core import cli as core_cli
 
 
 def create_parser():
@@ -26,12 +26,21 @@ def main():
 
 def prsint():
     """Console script for prsint."""
+    from .prsint import cli as prsint_cli
+
     parser = prsint_cli.prsint_parser()
     args = parser.parse_args()
 
-    print("Arguments: " + str(args))
-    # druint.test()
     prsint_cli.prsint(args)
+    return 0
+
+
+def add_variables():
+    """Console script for variable replacement."""
+    parser = core_cli.replace_parser()
+    args = parser.parse_args()
+
+    core_cli.replace_variables(args)
     return 0