+++ docs/whats_new.rst
@@ -17,6 +17,7 @@ New Features
 - Added :py:meth:`magic_numbers` to API (:pull:`85`).
 - New module for gfile creation from CMIP6 datasets (:pull:`78`, :pull:`89`).
 - New notebook on greenhouse gas concentration (GHG). This documents how we create GHG forcing tables for CMIP6 downscaling (:pull:`92`).
+- Command line interface ``pradd-vars`` for variable replacement: This tool can be used to replace soil variables in a forcing file with data from a REMO output file (*warm soil*) (:pull:`93`).
 
 Internal Changes
 ~~~~~~~~~~~~~~~~